Add files via upload
This is a Class B duplex in Gainesville that you plan to renovate before leasing out to students in order to command higher rents. You did your research and project the renovations to cost you $6,500 per unit in year 1. You can purchase the property for $178,500 today and you believe that the total combined rent will be $2,000 a month for the entire property with a 12 month lease. You also believe that your rent will grow at a constant yearly rate of 1.5% and project that you will have vacancy and collection losses of 3% annually since you secured creditable tenants. Both tenants will be charged $50 a month for their parking spot, maintenance and utilities will cost you $4,000 annually, and the operating expenses will grow at 2% yearly. Based on your assumptions you can sell your property in 5 years for $225,000. You expect there to be selling expenses of 6%. You will not invest in this property unless you can realize an unlevered before tax return of 10%, a levered before tax return of 14%, and a levered after tax return of 9.8%. You would like to acquire the property with 60% financed by ABC Bank with a 15 fixed interest rate loan at 6.5% per year. You also negotiate to only pay interest for the six months. You will have to pay 3% in loan expenses and you plan on using an amortization term of 30 years making this a fixed rate partially amortized loan with an interest only period. Your yearly taxes will be $3,750 and you will have taxes due on sale of 4% on the property.